Add test for single-case select without default
This change relates to golang/go#9966.
The added test warns of any select statements with a single case and no
default. Such a statement should be replaced with a simple channel
operation. The reason why a warning is appropriate is that the user may
have meant the make the issue non-blocking, but forgot the required
default case, potentially causing unexpected deadlock.